+++ examples/06_appview/app/views/app_view.js
@@ -4,7 +4,7 @@ var BaseAppView = require('rendr/client/app_view')
 ;
 
 module.exports = BaseAppView.extend({
-  postInitialize: function() {
+  initialize: function() {
     this.app.on('change:loading', function(app, loading) {
       $body.toggleClass('loading', loading);
     });