+++ shared/base/view.js
@@ -18,13 +18,19 @@ if (!isServer) {
 function noop() {}
 
 module.exports = BaseView = Backbone.View.extend({
-  initialize: function(options) {
+  constructor: function(options) {
+    this.cid = _.uniqueId('view');
+    this._configure(options || {});
+    this._ensureElement();
+
     var obj;
 
     this.parseOptions(options);
-
     this.name = this.name || this.app.modelUtils.underscorize(this.constructor.id || this.constructor.name);
-    this.postInitialize();
+
+    this.initialize.apply(this, arguments);
+    this.delegateEvents();
+
     if ((obj = this.model || this.collection) && this.renderOnRefresh) {
       obj.on('refresh', this.render, this);
     }
@@ -32,13 +38,18 @@ module.exports = BaseView = Backbone.View.extend({
     this.render = this.render.bind(this);
   },
 
+  initialize: function(options) {
+    if (this.postInitialize) {
+      console.warn('`postInitialize` is deprecated, please use `initialize`');
+      this.postInitialize();
+    }
+  },
+
   /**
    * Whether or not to re-render this view when the model or collection
    * emits a 'refresh' event. Used with 'model|collection.checkFresh()'.
    */
   renderOnRefresh: false,
-
-  postInitialize: noop,
 
   parseOptions: function(options) {
     /**